Added the lessfilter scripts to display objdump -x ... of ELF files.
Corrected the name of the create-packages.sh script (no more "-ve-").
Reorganized the tools & scripts in separate folders.
Added the new tools and scripts folder to the cp commandto create packages.
Depend on dh-exec and dh-systemd (see install-ubuntu-dependencies)
Use the dh-exec in the .install to rename the lessfilter as .lessfilter
Removed the auto-installed binaries from the .install file.