Adding script to restart pulseaudio.
On my Ubuntu 22.04, pulseaudio is not auto-restarted after a reboot.
The script makes sure to stop everything and then restarts the daemon.

The restart-system-monitor-extension is for Gnome users. Once in a while
(really rarely for me), the app. crashes and thus does not give me feedback
anymore. This script is useful to restart the app. without having to
remember all the commands.